*: new interface; use new github library
this integrates google/go-github instead of crosbymichael/octokat since
it no longer seems to be maintained. It also adds the pr grouping for
subcommands. issues coming hopefully soon which is why it was doen that
way.

Additionally, some UX improvements around using less for everything
(gutted), and some better feedback for watches; as well as some
concurrency around multiple prs for watch.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>